+++ gbe/models/profile.py
@@ -158,16 +158,12 @@ class Profile(WorkerItem):
                             reverse('profile_update',
                                     urlconf='gbe.urls'))
         for act in self.get_acts():
-            if act.accepted == 3 and \
-               act.is_current and \
-               (len(act.get_scheduled_rehearsals()) == 0 or
-                    not act.tech.is_complete):
+            if act.accepted == 3 and act.is_current and not act.is_complete:
                 p_alerts.append(
                     profile_alerts['schedule_rehearsal'] %
-                    (act.b_title,
-                     reverse('act_tech_wizard',
-                             urlconf='gbe.urls',
-                             args=[act.id])))
+                    (act.b_title, reverse('act_tech_wizard',
+                                          urlconf='gbe.urls',
+                                          args=[act.id])))
         return p_alerts
 
     def get_costumebids(self, historical=False):
