+++ scheduler/models/__init__.py
@@ -2,11 +2,11 @@ from .resource import Resource
 from .resource_item import ResourceItem
 from .location_item import LocationItem
 from .location import Location
+from .act_item import ActItem
+from .act_resource import ActResource
+from .schedulable import Schedulable
 from .remains import (
-    Schedulable,
     ResourceItem,
-    ActItem,
-    ActResource,
     WorkerItem,
     Worker,
     EventItem,
